Create res_vector.py
User inputs N points on a cartesian plane and the coordinate for each point (x, y) .
For each point, it computes the direction of the vector, the distance between the vectors and adds a unit force vector in that direction(Normalized by the distance.
Finally it then outputs the x and y components of the resultant force, the magnitude of the force and its angle with respect to the x axis.